refactor(types); extract public interfaces for VDatePicker
