+++ packages/vuetify/src/components/VDatePicker/VDatePickerHeader.ts
@@ -15,8 +15,8 @@ import mixins from '../../util/mixins'
 
 // Types
 import { VNode } from 'vue'
-import { DatePickerFormatter } from './util/createNativeLocaleFormatter'
 import { PropValidator } from 'vue/types/options'
+import { DatePickerFormatter } from './types'
 
 export default mixins(
   Colorable,
